Encapsulates specific Features for PWA - Separates HasServiceWorker & HasManifest.
Added methods updateServiceWorker() and getServiceWorker() to PwaManager.
