+++ gwt-material/src/main/java/gwt/material/design/client/pwa/manifest/HasManifest.java
@@ -1,0 +1,36 @@
+/*
+ * #%L
+ * GwtMaterial
+ * %%
+ * Copyright (C) 2015 - 2017 GwtMaterialDesign
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package gwt.material.design.client.pwa.manifest;
+
+public interface HasManifest {
+
+    /**
+     * Set up the manifest.json file to enable the Installable feature of GMD PWA
+     * @param manifestUrl - the url of your manifest.json file
+     */
+    void setupManifest(String manifestUrl);
+
+    /**
+     * Added a meta tag theme color to provide the browser's address
+     * bar a designated color by your app.
+     * @param themeColor - Hexadecimal value for your the background color of the address bar
+     */
+    void setupMetaThemeColor(String themeColor);
+}
